improved index.tsx
